Added New Success result code and move files in new folder
