+++ app/src/main/AndroidManifest.xml
@@ -27,7 +27,7 @@
             </intent-filter>
         </activity>
         
-        <receiver android:name=".utils.reminder.MyBroadcastReceiver" />
+        <receiver android:name=".common.reminder.MyBroadcastReceiver" />
     </application>
 
 </manifest>
