+++ app/src/main/java/org/example/moreinone/common/biometricAuthentication/BiometricResult.kt
@@ -1,4 +1,4 @@
-package org.example.moreinone.utils.biometricAuthentication
+package org.example.moreinone.common.biometricAuthentication
 
 sealed interface BiometricResult {
     data object HardwareUnavailable : BiometricResult
