+++ app/src/main/java/org/example/moreinone/common/reminder/MyBroadcastReceiver.kt
@@ -1,4 +1,4 @@
-package org.example.moreinone.utils.reminder
+package org.example.moreinone.common.reminder
 
 import android.content.BroadcastReceiver
 import android.content.Context
