+++ app/src/main/java/org/example/moreinone/common/reminder/NotificationHandler.kt
@@ -1,4 +1,4 @@
-package org.example.moreinone.utils.reminder
+package org.example.moreinone.common.reminder
 
 import android.app.NotificationChannel
 import android.app.NotificationManager
